# shared/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ecommerce-platform",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/storefront/orders": {
      "get": {
        "tags": [
          "storefront:orders"
        ],
        "summary": "List My Orders",
        "operationId": "list_my_orders_api_v1_storefront_orders_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderPageResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/storefront/orders/{order_id}": {
      "get": {
        "tags": [
          "storefront:orders"
        ],
        "summary": "Get My Order",
        "operationId": "get_my_order_api_v1_storefront_orders__order_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Order Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/storefront/orders/{order_id}/pay": {
      "post": {
        "tags": [
          "storefront:orders"
        ],
        "summary": "Pay Order",
        "operationId": "pay_order_api_v1_storefront_orders__order_id__pay_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Order Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PayOrderRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckoutUrlResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/admin/products/{product_id}/images/order": {
      "put": {
        "tags": [
          "admin:products"
        ],
        "summary": "Reorder Product Images",
        "operationId": "reorder_product_images_api_v1_admin_products__product_id__images_order_put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "product_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Product Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReorderImagesRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/orders": {
      "get": {
        "tags": [
          "admin:orders"
        ],
        "summary": "List Orders",
        "operationId": "list_orders_api_v1_admin_orders_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/OrderStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderPageResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/orders/{order_id}": {
      "get": {
        "tags": [
          "admin:orders"
        ],
        "summary": "Get Order",
        "operationId": "get_order_api_v1_admin_orders__order_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Order Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CheckoutUrlResponse": {
        "properties": {
          "checkout_url": {
            "type": "string",
            "title": "Checkout Url"
          }
        },
        "type": "object",
        "required": [
          "checkout_url"
        ],
        "title": "CheckoutUrlResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "OrderLineItemResponse": {
        "properties": {
          "product_id": {
            "type": "string",
            "format": "uuid",
            "title": "Product Id"
          },
          "product_name": {
            "type": "string",
            "title": "Product Name"
          },
          "unit_price": {
            "type": "string",
            "title": "Unit Price"
          },
          "quantity": {
            "type": "integer",
            "title": "Quantity"
          },
          "line_total": {
            "type": "string",
            "title": "Line Total"
          }
        },
        "type": "object",
        "required": [
          "product_id",
          "product_name",
          "unit_price",
          "quantity",
          "line_total"
        ],
        "title": "OrderLineItemResponse"
      },
      "OrderPageResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/OrderResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total"
        ],
        "title": "OrderPageResponse"
      },
      "OrderResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "customer_id": {
            "type": "string",
            "format": "uuid",
            "title": "Customer Id"
          },
          "status": {
            "$ref": "#/components/schemas/OrderStatus"
          },
          "payment_status": {
            "type": "string",
            "title": "Payment Status"
          },
          "line_items": {
            "items": {
              "$ref": "#/components/schemas/OrderLineItemResponse"
            },
            "type": "array",
            "title": "Line Items"
          },
          "subtotal": {
            "type": "string",
            "title": "Subtotal"
          },
          "tax": {
            "type": "string",
            "title": "Tax"
          },
          "shipping": {
            "type": "string",
            "title": "Shipping"
          },
          "total": {
            "type": "string",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "id",
          "customer_id",
          "status",
          "payment_status",
          "line_items",
          "subtotal",
          "tax",
          "shipping",
          "total"
        ],
        "title": "OrderResponse"
      },
      "OrderStatus": {
        "type": "string",
        "enum": [
          "pending",
          "paid",
          "fulfilled",
          "cancelled",
          "refunded"
        ],
        "title": "OrderStatus"
      },
      "PayOrderRequest": {
        "properties": {
          "success_url": {
            "type": "string",
            "minLength": 1,
            "title": "Success Url"
          },
          "cancel_url": {
            "type": "string",
            "minLength": 1,
            "title": "Cancel Url"
          }
        },
        "type": "object",
        "required": [
          "success_url",
          "cancel_url"
        ],
        "title": "PayOrderRequest"
      },
...
      "ProductResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "tenant_id": {
            "type": "string",
            "format": "uuid",
            "title": "Tenant Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "price": {
            "type": "string",
            "title": "Price"
          },
          "compare_at_price": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Compare At Price"
          },
          "sku": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sku"
          },
          "images": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Images"
          },
          "stock_qty": {
            "type": "integer",
            "title": "Stock Qty"
          },
          "status": {
            "$ref": "#/components/schemas/ProductStatus"
          },
          "category_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category Id"
          }
        },
        "type": "object",
        "required": [
          "id",
          "tenant_id",
          "name",
          "slug",
          "description",
          "price",
          "compare_at_price",
          "sku",
          "images",
          "stock_qty",
          "status",
          "category_id"
        ],
        "title": "ProductResponse"
      },
      "ProductStatus": {
        "type": "string",
        "enum": [
          "draft",
          "published"
        ],
        "title": "ProductStatus"
      },
...
      "ReorderImagesRequest": {
        "properties": {
          "ordered_urls": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Ordered Urls"
          }
        },
        "type": "object",
        "required": [
          "ordered_urls"
        ],
        "title": "ReorderImagesRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}